Organized parsers into separate modules within a 'parsers' folder. Image downloading logic is now integrated within the parse method itself.
