+++ parsers/parser_abc.py
@@ -1,0 +1,11 @@
+from abc import ABC, abstractmethod
+
+
+class Parser(ABC):
+
+    def __init__(self, url):
+        self.url = url
+
+    @abstractmethod
+    def parse(self):
+        pass
